convert user id from local return value from float64 to int

+++ src/helper/get-local.go
@@ -5,6 +5,6 @@ import (
 	"github.com/golang-jwt/jwt"
 )
 
-func GetUserIDFromLocals(c *fiber.Ctx) float64 {
-	return c.Locals("claims").(jwt.MapClaims)["id"].(float64)
+func GetUserIDFromLocals(c *fiber.Ctx) int {
+	return int(c.Locals("claims").(jwt.MapClaims)["id"].(float64))
 }
